+++ app/src/main/java/jp/co/tbdeveloper/warikanapp/feature_roulette/domain/repository/MemberFactory.kt
@@ -4,7 +4,7 @@ import jp.co.tbdeveloper.warikanapp.feature_roulette.domain.model.MemberEntity
 import jp.co.tbdeveloper.warikanapp.feature_roulette.domain.model.resource.Member
 
 object MemberFactory {
-    fun create(id: Int = 0, rouletteId: Int, members: List<Member>): List<MemberEntity> {
+    fun create(id: Long = 0, rouletteId: Long, members: List<Member>): List<MemberEntity> {
         return List(members.size) { i ->
             MemberEntity(id, rouletteId, members[i].name, members[i].color)
         }
