+++ app/src/main/java/jp/co/tbdeveloper/warikanapp/feature_roulette/domain/repository/RouletteFactory.kt
@@ -3,7 +3,7 @@ package jp.co.tbdeveloper.warikanapp.feature_roulette.domain.repository
 import jp.co.tbdeveloper.warikanapp.feature_roulette.domain.model.RouletteEntity
 
 object RouletteFactory {
-    fun create(id: Int = 0, total: Int): RouletteEntity {
+    fun create(id: Long = 0, total: Int): RouletteEntity {
         return RouletteEntity(id, total)
     }
 }
